assign parents to classes that are different from the classes they inherit from

+++ main/datatype/array_data/numeric_data/integer.py
@@ -1,5 +1,5 @@
 from main.datatype.array_data.numeric import Numeric
 
 
-class Float(Numeric):
+class Int(Numeric):
     pass
